+++ cmake/functions.cmake
@@ -1,6 +1,7 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright 2018-2022, Intel Corporation
+# Copyright (c) 2022 Fujitsu Limited
 #
 
 #
@@ -286,4 +287,21 @@ function(atomic_operations_supported var)
 		}"
 		ATOMIC_OPERATIONS_SUPPORTED)
 	set(var ${ATOMIC_OPERATIONS_SUPPORTED} PARENT_SCOPE)
+endfunction()
+
+# check if libibverbs has ibv_wr_atomic_write() support
+function(is_ibv_wr_atomic_write_supported var)
+	CHECK_C_SOURCE_COMPILES("
+		#include <infiniband/verbs.h>
+		/*
+		 * check if IB_UVERBS_DEVICE_ATOMIC_WRITE, IBV_QP_EX_WITH_ATOMIC_WRITE
+		 * and ibv_wr_atomic_write() are defined
+		 */
+		int main() {
+			uint64_t device_cap_flag = IB_UVERBS_DEVICE_ATOMIC_WRITE;
+			uint64_t send_ops_flag = IBV_QP_EX_WITH_ATOMIC_WRITE;
+			return !ibv_wr_atomic_write;
+		}"
+		IBV_WR_ATOMIC_WRITE_SUPPORTED)
+	set(var ${IBV_WR_ATOMIC_WRITE_SUPPORTED} PARENT_SCOPE)
 endfunction()