+++ tests/unit/common/mocks-rpma-utils.h
@@ -1,5 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /* Copyright 2020, Intel Corporation */
+/* Copyright (c) 2022, Fujitsu Limited */
 
 /*
  * mocks-rpma-utils.h -- the rpma-utils mocks' header
@@ -8,7 +9,9 @@
 #ifndef MOCKS_RPMA_UTILS_H
 #define MOCKS_RPMA_UTILS_H
 
-#define MOCK_ODP_CAPABLE	1
-#define MOCK_ODP_INCAPABLE	0
+#define MOCK_ODP_CAPABLE		1
+#define MOCK_ODP_INCAPABLE		0
+#define MOCK_ATOMIC_WRITE_CAPABLE	1
+#define MOCK_ATOMIC_WRITE_INCAPABLE	0
 
 #endif /* MOCKS_RPMA_UTILS_H */