+++ tests/unit/peer/peer-create_qp.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 /* Copyright 2020-2022, Intel Corporation */
-/* Copyright 2021-2022, Fujitsu */
+/* Copyright (c) 2021-2022, Fujitsu Limited */
 
 /*
  * peer-create_qp.c -- a peer unit test
@@ -40,7 +40,7 @@ static int num_rcqs = sizeof(rcqs) / sizeof(rcqs[0]);
  * configure_create_qp_ex -- configure common mock for rdma_create_qp_ex()
  */
 static void
-configure_create_qp_ex(struct rpma_cq *rcq)
+configure_create_qp_ex(struct prestate *prestate, struct rpma_cq *rcq)
 {
 	will_return(rpma_conn_cfg_get_sq_size, &Get_args);
 	will_return(rpma_conn_cfg_get_rq_size, &Get_args);
@@ -73,8 +73,17 @@ configure_create_qp_ex(struct rpma_cq *rcq)
 		RPMA_MAX_SGE);
 	expect_value(rdma_create_qp_ex, qp_init_attr->cap.max_inline_data,
 		RPMA_MAX_INLINE_DATA);
-	expect_value(rdma_create_qp_ex, qp_init_attr->comp_mask, IBV_QP_INIT_ATTR_PD);
 	expect_value(rdma_create_qp_ex, qp_init_attr->pd, MOCK_IBV_PD);
+#ifdef IBV_WR_ATOMIC_WRITE_SUPPORTED
+	if (prestate->is_atomic_write_capable) {
+		expect_value(rdma_create_qp_ex, qp_init_attr->comp_mask,
+				IBV_QP_INIT_ATTR_PD | IBV_QP_INIT_ATTR_SEND_OPS_FLAGS);
+		expect_value(rdma_create_qp_ex, qp_init_attr->send_ops_flags,
+				IBV_QP_EX_WITH_ATOMIC_WRITE);
+		return;
+	}
+#endif
+	expect_value(rdma_create_qp_ex, qp_init_attr->comp_mask, IBV_QP_INIT_ATTR_PD);
 }
 
 /*
@@ -134,7 +143,7 @@ create_qp__rdma_create_qp_ex_ERRNO(void **pprestate)
 	for (int i = 0; i < num_rcqs; i++) {
 		/* configure mock */
 		Get_args.srq = (i == 2) ? MOCK_RPMA_SRQ : NULL;
-		configure_create_qp_ex(rcqs[i]);
+		configure_create_qp_ex(prestate, rcqs[i]);
 		will_return(rdma_create_qp_ex, MOCK_ERRNO);
 
 		/* run test */
@@ -157,7 +166,7 @@ create_qp__success(void **pprestate)
 	for (int i = 0; i < num_rcqs; i++) {
 		/* configure mock */
 		Get_args.srq = (i == 2) ? MOCK_RPMA_SRQ : NULL;
-		configure_create_qp_ex(rcqs[i]);
+		configure_create_qp_ex(prestate, rcqs[i]);
 		will_return(rdma_create_qp_ex, MOCK_OK);
 
 		/* run test */
@@ -176,14 +185,14 @@ main(int argc, char *argv[])
 		/* rpma_peer_setup_qp() unit tests */
 		cmocka_unit_test(create_qp__peer_NULL),
 		cmocka_unit_test_prestate_setup_teardown(create_qp__id_NULL,
-				setup__peer, teardown__peer, &prestate_OdpCapable),
+				setup__peer, teardown__peer, &prestate_Capable),
 		cmocka_unit_test_prestate_setup_teardown(create_qp__cq_NULL,
-				setup__peer, teardown__peer, &prestate_OdpCapable),
+				setup__peer, teardown__peer, &prestate_Capable),
 		cmocka_unit_test_prestate_setup_teardown(
 				create_qp__rdma_create_qp_ex_ERRNO,
-				setup__peer, teardown__peer, &prestate_OdpCapable),
+				setup__peer, teardown__peer, &prestate_Capable),
 		cmocka_unit_test_prestate_setup_teardown(create_qp__success,
-				setup__peer, teardown__peer, &prestate_OdpCapable),
+				setup__peer, teardown__peer, &prestate_Capable),
 	};
 
 	return cmocka_run_group_tests(tests, NULL, NULL);