+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - offset of the beginning of the used persistent memory in the clients using PMem in the examples
 - one suppression for Memcheck on Ubuntu 22.04
 - CI Coverity build run once a day over the night
+- a check for the native atomic write support in libibverbs
+- internal APIs:
+  - rpma_utils_ibv_context_is_atomic_write_capable() - checks if kernel supports native atomic write
 
 ### Fixed
 - DEVELOPMENT.md file - `CMAKE_BUILD_TYPE` must be set to `Debug` when running the tests
@@ -38,6 +41,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the default 'master' branch has been renamed to 'main'
 - logging of the source and the destination GID addresses in rpma_conn_req_new_from_id()
   has been restricted to only one case when CMAKE_BUILD_TYPE is set to 'Debug'
+- rpma_peer_new() to check the native atomic write support in kernel
+- rpma_peer_setup_qp() to enable native atomic write if both kernel and libibverbs supported it
+- rpma_mr_atomic_write() to use native atomic write if the created QP supported it
 
 ### Removed
 - whole benchmarking framework for librpma (the last commit with the benchmarking framework present is marked with the "[benchmarking-framework][bench-frame]" tag)